Fixing InventoryItemWidget OnClick not firing because it was not declared as a UFUNCTION.
Switching InventorySlot back to using pure UInventoryObject
pointers instead of TSoftObjectPtr because the TSoftObjectPtr
was not ensuring that the referenced UInventoryObject was actually
loaded when trying to add it to the inventory.

Future changes might alter this in some way to allow for delayed
loading, though UInventoryObjects are very lightweight so far.